+++ src/main/groovy/com/tkruse/gradle/groovysh/BuildShellTaskExtension.groovy
@@ -5,6 +5,6 @@ package com.tkruse.gradle.groovysh
  */
 class BuildShellTaskExtension extends ShellTaskExtension {
 
-    String gradleVersion = '1.12'
+    String gradleVersion = '2.0'
 
 }