+++ src/main/groovy/com/tkruse/gradle/groovysh/GroovyshPlugin.groovy
@@ -4,6 +4,7 @@ import org.gradle.api.GradleScriptException
 import org.gradle.api.InvalidUserDataException
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.util.GradleVersion
 
 /**
  * Plugin main class
@@ -39,8 +40,15 @@ class GroovyshPlugin implements Plugin<Project> {
                 throw new GradleScriptException("$NAME: Cannot create task ${BuildDevShellTask.NAME}", e)
             }
             URLClassLoader loader = GroovyObject.classLoader
-            // groovy < 2.2.0 groovysh runs on jline 1.0
-            project.dependencies.add(BuildDevShellTask.CONFIGURATION_NAME, 'jline:jline:1.0')
+
+            if (GradleVersion.current().getVersion().startsWith('1')) {
+                // gradle < 2.0 runs with groovy 1.8.6, groovysh needs jline 1.0
+                project.dependencies.add(BuildDevShellTask.CONFIGURATION_NAME, 'jline:jline:1.0')
+            } else {
+                // gradle 2.0 runs with groovy 2.3.3, groovysh needs jline 2.11
+                project.dependencies.add(BuildDevShellTask.CONFIGURATION_NAME, 'jline:jline:2.11')
+            }
+
             project.configurations.getByName(BuildDevShellTask.CONFIGURATION_NAME).each { File file ->
                 loader.addURL(file.toURL())
             }